back to functional point
Added a "util" module in order to organize the mud a bit more. Went ahead and made some adjustments to commands and utils so that I no longer need to manually import all the things but instead handle it a bit more cleanly.

Timing issue exists with util vs commands and I opted to temporarily put util loading early in the script to prevent this timing issue, or get stalled in its remediation.